refactor: shuffle steps around to make room for 1b
which will hold functionality to list and create koopovereenkomst. Will become step2
