+++ koopovereenkomst-v3-simple/pages/steps_verkoper/Step1.tsx
@@ -16,9 +16,9 @@ import ConnectSolid from "../../src/ConnectSolid";
 import { checkIfWebIDIsReady, registerWebID } from "../../src/mosService";
 
 import { VCARD } from "@inrupt/vocab-common-rdf";
-import { SolidPerson, SolidAddress } from "../../src/SolidInterfaces";
+import { SolidPerson, SolidAddress } from "../../src/Solid";
 
-export default function Step1({ handleNext, handleBack = () => {} }) {
+export function Step1({ handleNext, handleBack = () => {} }) {
   const { session } = useSession();
 
   const [isReady, setIsReady] = useState(null as boolean);
