+++ koopovereenkomst-v3-simple/pages/steps_verkoper/Step5.tsx
@@ -4,7 +4,7 @@ import Stack from "@mui/material/Stack";
 import Button from "@mui/material/Button";
 import { Box } from "@mui/system";
 
-export default function Step5({ step = 5, handleNext, handleBack = () => {} }) {
+export function Step5({ step = 5, handleNext, handleBack = () => {} }) {
   return (
     <Box sx={{ flex: 1 }}>
       <Typography variant="h1" color="text.primary" align="center">
