+++ koopovereenkomst-v3-simple/pages/steps_verkoper/Step6.tsx
@@ -4,7 +4,7 @@ import Stack from "@mui/material/Stack";
 import Button from "@mui/material/Button";
 import { Box } from "@mui/system";
 
-export default function Step6({ step = 6, handleBack = () => {} }) {
+export function Step6({ step = 6, handleBack = () => {} }) {
   return (
     <Box sx={{ flex: 1 }}>
       <Typography variant="h1" color="text.primary" align="center">
