+++ koopovereenkomst-v3-simple/pages/steps_verkoper/index.ts
@@ -1,0 +1,7 @@
+export { Step1 } from "./Step1";
+export { Step1b } from "./Step1b";
+export { Step2 } from "./Step2";
+export { Step3 } from "./Step3";
+export { Step4 } from "./Step4";
+export { Step5 } from "./Step5";
+export { Step6 } from "./Step6";
